# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bank Rest API
  version: 1.0.0

paths:
  /api/v1/auth/registration:
    post:
      description: Создание нового пользователя
      operationId: registration
      tags:
        - authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrationRequest'
      responses:
        '201':
          description: Пользователь создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenDto"
        '400':
          description: Неправильный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/auth/login:
    post:
      description: Авторизация пользователя
      operationId: login
      tags:
        - authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Пользователь авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenDto"
        '400':
          description: Неверный логин или пароль
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/card/my:
    get:
      description: Просмотр своих карт (поиск + пагинация)
      operationId: getMyCards
      tags:
        - card
      x-spring-paginated: true
      parameters:
        - name: search
          in: query
          description: "Строка поиска"
          schema:
            type: string
      responses:
        '200':
          description: Карты пользователя с пагинацией
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageResponseCardDto"
        '400':
          description: Неверные данные
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/card/all:
    get:
      description: Просмотр своих карт (поиск + пагинация) (АДМИН)
      operationId: getAllCards
      tags:
        - card
      x-spring-paginated: true
      parameters:
        - name: search
          in: query
          description: "Строка поиска"
          schema:
            type: string
      responses:
        '200':
          description: Карты пользователей с пагинацией
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageResponseCardDto"
        '400':
          description: Неверные данные
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/card/{cardId}/blocked-request:
    post:
      description: Создание заявки на блокировку
      operationId: requestBlock
      tags:
        - cardBlocked
      parameters:
        - name: cardId
          required: true
          in: path
          description: "Уникальный идентификатор карты"
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Номер заявки
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardBlockedResponseDto"
        '400':
          description: Неверные данные
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/card/all/blocked-request:
    get:
      description: Получить список заявок на блокировку
      operationId: getAllBlockedRequest
      tags:
        - cardBlocked
      x-spring-paginated: true
      responses:
        '200':
          description: Список заявок
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageCardBlockedRequestDto"
        '400':
          description: Неверные данные
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/card/blocked-request:
    post:
      description: Заблокировать или отклонить блокировку карты
      operationId: processBlockRequest
      tags:
        - cardBlocked
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessBlockDto"
      responses:
        '200':
          description: Информация о решение об блокировки карты
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessBlockDto"
        '400':
          description: Неверные данные
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/card/translation:
    post:
      description: Перевод между своими картами
      operationId: translation
      tags:
        - card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TranslationRequest'
      responses:
        '200':
          description: Информация о переводе
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranslationResponse"
        '400':
          description: Неверные данные
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/card/status:
    post:
      description: Блокировка пользователя
      operationId: cardStatus
      tags:
        - card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardStatusDto'
      responses:
        '200':
          description: Информация о блокировки
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseId"
        '400':
          description: Неверные данные
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/card/{cardId}/delete:
    post:
      description: Блокировка пользователя
      operationId: cardDelete
      tags:
        - card
      parameters:
        - name: cardId
          required: true
          in: path
          description: "Уникальный идентификатор карты"
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Информация об удалении
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseId"
        '400':
          description: Неверные данные
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/card/create:
    post:
      description: Создание карты для пользователя
      operationId: cardCreate
      tags:
        - card
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCardDto"
      responses:
        '200':
          description: Информация созданной карты
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardDto"
        '400':
          description: Неверные данные
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/user/blocked:
    post:
      description: Блокировка пользователя
      operationId: userBlocked
      tags:
        - user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserBlockedDto'
      responses:
        '200':
          description: Информация о блокировки
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseId"
        '400':
          description: Неверные данные
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        code:
          type: integer

    RegistrationRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        password:
          type: string

    LoginRequest:
      type: object
      properties:
        login:
          type: string
        password:
          type: string

    TokenDto:
      type: object
      properties:
        jwt:
          type: string

    CardBlockedResponseDto:
      type: object
      properties:
        applicationNumber:
          type: string

    CardBlockedRequestDto:
      type: object
      properties:
        number:
          type: string
          format: uuid
        cardId:
          type: integer
          format: int64
        statusCard:
          $ref: "#/components/schemas/StatusCardBlockedEnum"
        createDate:
          type: string
          format: date-time

    PageCardBlockedRequestDto:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/CardBlockedRequestDto'

    PageMetadata:
      type: object
      properties:
        page:
          type: integer
        size:
          type: integer
        totalPages:
          type: integer
        totalElements:
          type: integer
          format: int64

    PageResponseCardDto:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/CardDto'

    CardDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 123
        numberCard:
          type: string
          example: "**** **** **** 1234"
        fullName:
          type: string
          example: "Ivan Ivanov"
        validityDate:
          type: string
          example: "12/12"
          description: "Срок карты"
        status:
          $ref: '#/components/schemas/StatusCardEnum'
        balance:
          type: number
          format: double
          example: 1000.50
      required:
        - id
        - numberCard
        - fullName
        - validityDate
        - status
        - balance

    CreateCardDto:
      type: object
      properties:
        UserId:
          type: integer
          format: int64

    TranslationResponse:
      type: object
      properties:
        fromCard:
          type: string
          description: Номер карты с которой будет перевод
        toCard:
          description: Номер карты на которую будет зачисление
          type: string
        count:
          description: Сумма
          type: string
          format: decimal

    TranslationRequest:
      type: object
      properties:
        fromCard:
          type: string
          description: Номер карты с которой был перевод
        toCard:
          description: Номер карты на которую было зачисление
          type: string
        count:
          type: string
          format: decimal
        requestId:
          type: string
          format: uuid

    ResponseId:
      type: object
      properties:
        id:
          type: integer
          format: int64

    UserBlockedDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        block:
          type: boolean

    CardStatusDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        cardStatus:
          $ref: "#/components/schemas/StatusCardEnum"

    ProcessBlockDto:
      type: object
      properties:
        applicationNumber:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/StatusCardBlockedEnum"

    StatusCardEnum:
      type: string
      enum:
        - ACTION
        - BLOCKED
        - EXPIRED

    StatusCardBlockedEnum:
      type: string
      enum:
        - PENDING
        - APPROVED
        - REJECTED

  parameters:
    QueryRequest:
      name: queryRequest
      in: query
      description: "Параметры запроса для фильтрации и пагинации"
      schema:
        type: object
        properties:
          page:
            type: integer
            default: 0
          size:
            type: integer
            default: 20
          sort:
            type: array
            items:
              type: string
            description: "Сортировка: field,asc или field,desc"
          search:
            type: string